@@ -1,29 +1,6 @@
-/*
- * MIT License
- *
- * Copyright (c) 2020-2023 Jakub Zagórski (jaqobb)
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all
- * copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
- */
-
 package dev.jaqobb.rewardable_activities.listener.entity;
 
+import com.cryptomorin.xseries.XEntityType;
 import dev.jaqobb.rewardable_activities.RewardableActivitiesConstants;
 import dev.jaqobb.rewardable_activities.RewardableActivitiesPlugin;
 import dev.jaqobb.rewardable_activities.data.RewardableActivity;
@@ -38,20 +15,19 @@ import org.bukkit.event.Listener;
 import org.bukkit.event.entity.EntityDamageByEntityEvent;
 import org.bukkit.projectiles.ProjectileSource;
 
-public final class EntityDamageByEntityListener implements Listener {
-
+public class EntityDamageByEntityListener implements Listener {
+    
     private final RewardableActivitiesPlugin plugin;
-
+    
     public EntityDamageByEntityListener(RewardableActivitiesPlugin plugin) {
         this.plugin = plugin;
     }
-
+    
     @EventHandler(priority = EventPriority.MONITOR)
     public void onEntityDamageByEntity(EntityDamageByEntityEvent event) {
-        if (!(event.getEntity() instanceof LivingEntity)) {
+        if (!(event.getEntity() instanceof LivingEntity victim)) {
             return;
         }
-        LivingEntity victim = (LivingEntity) event.getEntity();
         Player attacker = this.getPlayer(event.getDamager());
         if (attacker == null) {
             return;
@@ -65,26 +41,27 @@ public final class EntityDamageByEntityListener implements Listener {
         if (this.plugin.isEntitySpawnerOwnershipCheckEnabled() && this.plugin.hasMetadata(victim, RewardableActivitiesConstants.ENTITY_SPAWNED_BY_SPAWNER_KEY)) {
             return;
         }
-        RewardableActivity activity = this.plugin.getRepository().getEntityKillActivity(victim.getType());
+        RewardableActivity activity = this.plugin.getRepository().getEntityKillActivity(XEntityType.of(victim));
         if (activity == null) {
             return;
         }
         RewardableActivityReward reward = activity.getReward(attacker);
-        if (reward != null) {
-            reward.reward(this.plugin, attacker);
+        if (reward == null) {
+            return;
         }
+        reward.reward(this.plugin, attacker);
     }
-
+    
     private Player getPlayer(Entity entity) {
         if (entity instanceof Player) {
             return (Player) entity;
         }
-        if (entity instanceof Projectile) {
-            Projectile projectile = (Projectile) entity;
+        if (entity instanceof Projectile projectile) {
             ProjectileSource projectileSource = projectile.getShooter();
-            if (projectileSource instanceof Player) {
-                return (Player) projectileSource;
+            if (!(projectileSource instanceof Player)) {
+                return null;
             }
+            return (Player) projectileSource;
         }
         return null;
     }
