@@ -1,49 +1,26 @@
-/*
- * MIT License
- *
- * Copyright (c) 2020-2023 Jakub Zagórski (jaqobb)
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all
- * copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
- */
-
 package dev.jaqobb.rewardable_activities.updater;
 
 import dev.jaqobb.rewardable_activities.RewardableActivitiesConstants;
 import dev.jaqobb.rewardable_activities.RewardableActivitiesPlugin;
+import net.md_5.bungee.api.ChatColor;
+import org.bukkit.scheduler.BukkitRunnable;
 import java.io.BufferedReader;
 import java.io.InputStream;
 import java.io.InputStreamReader;
+import java.net.URI;
 import java.net.URL;
 import java.nio.charset.StandardCharsets;
 import java.util.logging.Level;
 import javax.net.ssl.HttpsURLConnection;
-import net.md_5.bungee.api.ChatColor;
-import org.bukkit.scheduler.BukkitRunnable;
 
-public final class Updater extends BukkitRunnable {
-
+public class Updater extends BukkitRunnable {
+    
     private final RewardableActivitiesPlugin plugin;
     private final int pluginId;
     private final String currentVersion;
     private String latestVersion;
     private Integer versionDifference;
-
+    
     public Updater(RewardableActivitiesPlugin plugin, int pluginId) {
         this.plugin = plugin;
         this.pluginId = pluginId;
@@ -51,62 +28,64 @@ public final class Updater extends BukkitRunnable {
         this.latestVersion = null;
         this.versionDifference = null;
     }
-
+    
     public String getCurrentVersion() {
         return this.currentVersion;
     }
-
+    
     public String getLatestVersion() {
         return this.latestVersion;
     }
-
+    
     public Integer getVersionDifference() {
         return this.versionDifference;
     }
-
+    
     public String getUpdateMessage() {
-        String message = RewardableActivitiesConstants.PREFIX;
         if (this.currentVersion.contains("-SNAPSHOT")) {
-            message += ChatColor.RED + "You are running a development version (" + ChatColor.WHITE + this.currentVersion + ChatColor.RED + "). It is not advised to run development versions on production servers as they are very likely to not work as intended.";
-        } else if (this.latestVersion == null || this.versionDifference == null) {
-            message += ChatColor.RED + "Could not retrieve the latest version data. Make sure that you have internet access.";
-        } else if (this.versionDifference > 0) {
-            message += ChatColor.WHITE + "You are running a future version (" + ChatColor.GRAY + this.currentVersion + ChatColor.WHITE + " > " + ChatColor.GRAY + this.latestVersion + ChatColor.WHITE + "). This version is safe to use but is yet to be officially uploaded or the latest version data is yet to be updated.";
-        } else if (this.versionDifference < 0) {
-            message += ChatColor.WHITE + "You are running a past version (" + ChatColor.GRAY + this.currentVersion + ChatColor.WHITE + " < " + ChatColor.GRAY + this.latestVersion + ChatColor.WHITE + "). Updating is recommended to receive new features, bug fixes and more.";
-        } else {
-            message += ChatColor.WHITE + "You are running the latest version (" + ChatColor.GRAY + this.latestVersion + ChatColor.WHITE + ").";
+            return RewardableActivitiesConstants.PREFIX + ChatColor.RED + "You are running a development version of " + ChatColor.WHITE + "Rewardable Activities" + ChatColor.RED + " (" + ChatColor.WHITE + this.currentVersion + ChatColor.RED + "). Development versions may be unstable. As such, please avoid running them on production servers.";
         }
-        return message;
+        if (this.latestVersion == null || this.versionDifference == null) {
+            return RewardableActivitiesConstants.PREFIX + ChatColor.RED + "Could not retrieve the latest version data of " + ChatColor.WHITE + "Rewardable Activities" + ChatColor.RED + ".";
+        }
+        if (this.versionDifference < 0) {
+            return RewardableActivitiesConstants.PREFIX + ChatColor.WHITE + "You are running an outdated version of " + ChatColor.GRAY + "Rewardable Activities" + ChatColor.WHITE + " (" + ChatColor.GRAY + this.currentVersion + ChatColor.WHITE + " < " + ChatColor.GRAY + this.latestVersion + ChatColor.WHITE + "). Consider updating to receive new features, bug fixes, and more.";
+        }
+        if (this.versionDifference > 0) {
+            return RewardableActivitiesConstants.PREFIX + ChatColor.WHITE + "You are running a future version of " + ChatColor.GRAY + "Rewardable Activities" + ChatColor.WHITE + " (" + ChatColor.GRAY + this.currentVersion + ChatColor.WHITE + " > " + ChatColor.GRAY + this.latestVersion + ChatColor.WHITE + "). I suppose you are a time traveler.";
+        }
+        return RewardableActivitiesConstants.PREFIX + ChatColor.WHITE + "You are running the latest version of " + ChatColor.GRAY + "Rewardable Activities" + ChatColor.WHITE + ".";
     }
-
+    
     @Override
     public void run() {
         if (this.currentVersion.contains("-SNAPSHOT")) {
             return;
         }
         try {
-            HttpsURLConnection connection = (HttpsURLConnection) new URL("https://api.spigotmc.org/legacy/update.php?resource=" + this.pluginId).openConnection();
+            HttpsURLConnection connection = (HttpsURLConnection) URL.of(URI.create("https://api.spigotmc.org/legacy/update.php?resource=" + this.pluginId), null).openConnection();
             connection.setRequestMethod("GET");
             try (InputStream input = connection.getInputStream(); BufferedReader reader = new BufferedReader(new InputStreamReader(input, StandardCharsets.UTF_8))) {
                 this.latestVersion = reader.readLine();
                 String[] currentVersionData = this.currentVersion.split("\\.");
                 String[] latestVersionData = this.latestVersion.split("\\.");
-                if (currentVersionData.length == 3 && latestVersionData.length == 3) {
-                    int majorDifference = Integer.compare(Integer.parseInt(currentVersionData[0]), Integer.parseInt(latestVersionData[0]));
-                    if (majorDifference != 0) {
-                        this.versionDifference = majorDifference;
-                    } else {
-                        int minorDifference = Integer.compare(Integer.parseInt(currentVersionData[1]), Integer.parseInt(latestVersionData[1]));
-                        if (minorDifference != 0) {
-                            this.versionDifference = minorDifference;
-                        } else {
-                            this.versionDifference = Integer.compare(Integer.parseInt(currentVersionData[2]), Integer.parseInt(latestVersionData[2]));
-                        }
-                    }
+                if (currentVersionData.length != 3 || latestVersionData.length != 3) {
+                    return;
                 }
+                int majorDifference = Integer.compare(Integer.parseInt(currentVersionData[0]), Integer.parseInt(latestVersionData[0]));
+                if (majorDifference != 0) {
+                    this.versionDifference = majorDifference;
+                    return;
+                }
+                int minorDifference = Integer.compare(Integer.parseInt(currentVersionData[1]), Integer.parseInt(latestVersionData[1]));
+                if (minorDifference != 0) {
+                    this.versionDifference = minorDifference;
+                    return;
+                }
+                this.versionDifference = Integer.compare(Integer.parseInt(currentVersionData[2]), Integer.parseInt(latestVersionData[2]));
+            } finally {
+                connection.disconnect();
             }
-            connection.disconnect();
         } catch (Exception exception) {
             this.plugin.getLogger().log(Level.WARNING, "Could not retrieve the latest version data.", exception);
         }
