@@ -1,27 +1,3 @@
-/*
- * MIT License
- *
- * Copyright (c) 2020-2023 Jakub Zagórski (jaqobb)
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all
- * copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
- */
-
 package dev.jaqobb.rewardable_activities;
 
 import dev.jaqobb.rewardable_activities.command.RewardableActivitiesCommand;
@@ -41,11 +17,7 @@ import dev.jaqobb.rewardable_activities.listener.player.PlayerJoinListener;
 import dev.jaqobb.rewardable_activities.listener.plugin.PluginDisableListener;
 import dev.jaqobb.rewardable_activities.listener.plugin.PluginEnableListener;
 import dev.jaqobb.rewardable_activities.updater.Updater;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.logging.Level;
 import net.milkbowl.vault.economy.Economy;
-import org.bstats.bukkit.Metrics;
 import org.bukkit.block.Block;
 import org.bukkit.block.BlockFace;
 import org.bukkit.metadata.FixedMetadataValue;
@@ -53,32 +25,32 @@ import org.bukkit.metadata.Metadatable;
 import org.bukkit.plugin.PluginManager;
 import org.bukkit.plugin.RegisteredServiceProvider;
 import org.bukkit.plugin.java.JavaPlugin;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.logging.Level;
 
-public final class RewardableActivitiesPlugin extends JavaPlugin {
-
+public class RewardableActivitiesPlugin extends JavaPlugin {
+    
     private boolean blockBreakOwnershipCheckEnabled;
     private boolean blockPlaceOwnershipCheckEnabled;
     private boolean entityBreedOwnershipCheckEnabled;
     private boolean entitySpawnerOwnershipCheckEnabled;
     private RewardableActivityRepository repository;
     private boolean placeholderApiPresent;
-    private Metrics metrics;
     private Updater updater;
     private Economy economy;
-
+    
     @Override
     public void onLoad() {
         this.saveDefaultConfig();
         this.loadConfig(false);
         PluginManager pluginManager = this.getServer().getPluginManager();
-        this.placeholderApiPresent = pluginManager.getPlugin(RewardableActivitiesConstants.PLACEHOLDER_API_PLUGIN_NAME) != null;
-        this.getLogger().log(Level.INFO, RewardableActivitiesConstants.PLACEHOLDER_API_PLUGIN_NAME + ": " + (this.placeholderApiPresent ? "present" : "not present") + ".");
+        this.placeholderApiPresent = pluginManager.getPlugin("PlaceholderAPI") != null;
+        this.getLogger().log(Level.INFO, "PlaceholderAPI: " + (this.placeholderApiPresent ? "present" : "not present") + ".");
     }
-
+    
     @Override
     public void onEnable() {
-        this.getLogger().log(Level.INFO, "Starting metrics...");
-        this.metrics = new Metrics(this, 9499);
         this.getLogger().log(Level.INFO, "Starting updater...");
         this.updater = new Updater(this, 86090);
         this.updater.runTaskTimerAsynchronously(this, 0L, 20L * 60L * 30L);
@@ -108,7 +80,7 @@ public final class RewardableActivitiesPlugin extends JavaPlugin {
         pluginManager.registerEvents(new PluginDisableListener(this), this);
         pluginManager.registerEvents(new PluginEnableListener(this), this);
     }
-
+    
     public void loadConfig(boolean reload) {
         this.getLogger().log(Level.INFO, (reload ? "Rel" : "L") + "oading configuration...");
         this.blockBreakOwnershipCheckEnabled = this.getConfig().getBoolean("block.ownership-check.break", this.getConfig().getBoolean("block.ownership-check.place", this.getConfig().getBoolean("block.ownership-check")));
@@ -127,63 +99,59 @@ public final class RewardableActivitiesPlugin extends JavaPlugin {
         this.getLogger().log(Level.INFO, " * Entity breed: " + this.repository.getEntityBreedActivities().size());
         this.getLogger().log(Level.INFO, " * Item fish: " + this.repository.getItemFishActivities().size());
     }
-
+    
     public boolean isBlockBreakOwnershipCheckEnabled() {
         return this.blockBreakOwnershipCheckEnabled;
     }
-
+    
     public boolean isBlockPlaceOwnershipCheckEnabled() {
         return this.blockPlaceOwnershipCheckEnabled;
     }
-
+    
     public boolean isEntityBreedOwnershipCheckEnabled() {
         return this.entityBreedOwnershipCheckEnabled;
     }
-
+    
     public boolean isEntitySpawnerOwnershipCheckEnabled() {
         return this.entitySpawnerOwnershipCheckEnabled;
     }
-
+    
     public RewardableActivityRepository getRepository() {
         return this.repository;
     }
-
+    
     public boolean isPlaceholderApiPresent() {
         return this.placeholderApiPresent;
     }
-
+    
     public void setPlaceholderApiPresent(boolean present) {
         this.placeholderApiPresent = present;
     }
-
-    public Metrics getMetrics() {
-        return this.metrics;
-    }
-
+    
     public Updater getUpdater() {
         return this.updater;
     }
-
+    
     public Economy getEconomy() {
         return this.economy;
     }
-
+    
     public boolean hasMetadata(Metadatable metadatable, String key) {
         return metadatable.hasMetadata(key);
     }
-
+    
     public void setMetadata(Metadatable metadatable, String key, Object value) {
         if (!metadatable.hasMetadata(key)) {
             metadatable.setMetadata(key, new FixedMetadataValue(this, value));
         }
     }
-
+    
     public void unsetMetadata(Metadatable metadatable, String key) {
         if (metadatable.hasMetadata(key)) {
             metadatable.removeMetadata(key, this);
         }
     }
-
+    
     public void updatePistonBlocks(BlockFace direction, List<Block> blocks) {
         List<Block> blocksPlacedByPlayer = new ArrayList<>(blocks.size());
         List<Block> blocksSoonToBePlacedByPlayer = new ArrayList<>(blocks.size());
@@ -203,7 +171,7 @@ public final class RewardableActivitiesPlugin extends JavaPlugin {
             }
         }
     }
-
+    
     private Economy setupEconomy() {
         if (!this.getServer().getPluginManager().isPluginEnabled("Vault")) {
             return null;
