@@ -1,27 +1,3 @@
-/*
- * MIT License
- *
- * Copyright (c) 2020-2023 Jakub Zagórski (jaqobb)
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all
- * copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
- */
-
 package dev.jaqobb.rewardable_activities.listener.block;
 
 import com.cryptomorin.xseries.XMaterial;
@@ -36,27 +12,23 @@ import org.bukkit.event.EventPriority;
 import org.bukkit.event.Listener;
 import org.bukkit.event.block.BlockPlaceEvent;
 
-public final class BlockPlaceListener implements Listener {
-
+public class BlockPlaceListener implements Listener {
+    
     private final RewardableActivitiesPlugin plugin;
-
+    
     public BlockPlaceListener(RewardableActivitiesPlugin plugin) {
         this.plugin = plugin;
     }
-
+    
     @EventHandler(priority = EventPriority.MONITOR, ignoreCancelled = true)
     public void onBlockPlace(BlockPlaceEvent event) {
         Block block = event.getBlock();
         if (this.plugin.isBlockBreakOwnershipCheckEnabled()) {
             this.plugin.setMetadata(block, RewardableActivitiesConstants.BLOCK_PLACED_BY_PLAYER_KEY, true);
         }
-        boolean placedByPlayer = this.plugin.hasMetadata(block, RewardableActivitiesConstants.BLOCK_PLACED_BY_PLAYER_2_KEY);
-        if (this.plugin.isBlockPlaceOwnershipCheckEnabled()) {
-            this.plugin.setMetadata(block, RewardableActivitiesConstants.BLOCK_PLACED_BY_PLAYER_2_KEY, true);
-            if (placedByPlayer || this.plugin.hasMetadata(block, RewardableActivitiesConstants.BLOCK_BROKEN_BY_PLAYER_KEY)) {
-                this.plugin.unsetMetadata(block, RewardableActivitiesConstants.BLOCK_BROKEN_BY_PLAYER_KEY);
-                return;
-            }
+        if (this.plugin.isBlockPlaceOwnershipCheckEnabled() && this.plugin.hasMetadata(block, RewardableActivitiesConstants.BLOCK_BROKEN_BY_PLAYER_KEY)) {
+            this.plugin.unsetMetadata(block, RewardableActivitiesConstants.BLOCK_BROKEN_BY_PLAYER_KEY);
+            return;
         }
         RewardableActivity activity = this.plugin.getRepository().getBlockPlaceActivity(XMaterial.matchXMaterial(block.getType()));
         if (activity == null) {
@@ -64,8 +36,9 @@ public final class BlockPlaceListener implements Listener {
         }
         Player player = event.getPlayer();
         RewardableActivityReward reward = activity.getReward(player);
-        if (reward != null) {
-            reward.reward(this.plugin, player);
+        if (reward == null) {
+            return;
         }
+        reward.reward(this.plugin, player);
     }
 }
