@@ -1,27 +1,3 @@
-/*
- * MIT License
- *
- * Copyright (c) 2020-2023 Jakub Zagórski (jaqobb)
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all
- * copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
- */
-
 package dev.jaqobb.rewardable_activities.listener.entity;
 
 import dev.jaqobb.rewardable_activities.RewardableActivitiesConstants;
@@ -31,18 +7,19 @@ import org.bukkit.event.EventPriority;
 import org.bukkit.event.Listener;
 import org.bukkit.event.entity.SpawnerSpawnEvent;
 
-public final class SpawnerSpawnListener implements Listener {
-
+public class SpawnerSpawnListener implements Listener {
+    
     private final RewardableActivitiesPlugin plugin;
-
+    
     public SpawnerSpawnListener(RewardableActivitiesPlugin plugin) {
         this.plugin = plugin;
     }
-
+    
     @EventHandler(priority = EventPriority.MONITOR)
     public void onEntitySpawn(SpawnerSpawnEvent event) {
-        if (this.plugin.isEntitySpawnerOwnershipCheckEnabled()) {
-            this.plugin.setMetadata(event.getEntity(), RewardableActivitiesConstants.ENTITY_SPAWNED_BY_SPAWNER_KEY, true);
+        if (!this.plugin.isEntitySpawnerOwnershipCheckEnabled()) {
+            return;
         }
+        this.plugin.setMetadata(event.getEntity(), RewardableActivitiesConstants.ENTITY_SPAWNED_BY_SPAWNER_KEY, true);
     }
 }
